internal/common/awsxray/util/util.go
// Copyright The OpenTelemetry Authors
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

package util

// String returns a pointer to the string value passed in.
func String(v string) *string {
	return &v
}

// Bool returns a pointer to the bool value passed in.
func Bool(v bool) *bool {
	return &v
}

// Int64 returns a pointer to the bool value passed in.
func Int64(v int64) *int64 {
	return &v
}

// Int returns a pointer to the bool value passed in.
func Int(v int) *int {
	return &v
}

// Float64 returns a pointer to the bool value passed in.
func Float64(v float64) *float64 {
	return &v
}
